requirements.txt
attrs==18.2.0
cattrs==0.9.0
Flask==1.0.2
flask-swagger-ui==3.18.0
httplib2==0.9.1
MarkupSafe==0.23
oauthlib==1.0.3
Pillow==3.1.2
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pyparsing==2.0.3
pyxdg==0.25
PyYAML==3.13
reportlab==3.3.0
requests==2.9.1
six==1.10.0
typing==3.6.6
urllib3==1.13.1
Werkzeug==0.14.1
gunicorn==19.9.0
